requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.3
djangorestframework==3.15.2
dnspython==2.7.0
pymongo==4.10.1
sqlparse==0.5.2
